feat: add assignment 4.21

+++ part4/blog-list/controllers/blogs.js
@@ -46,8 +46,22 @@ blogsRouter.get('/:id', async (req, res) => {
 });
 
 blogsRouter.delete('/:id', async (request, response) => {
-	const blog = await Blog.findByIdAndRemove(request.params.id);
-	response.status(204).end();
+	const decodedToken = jwt.verify(request.token, process.env.SECRET);
+
+	if (!request.token || !decodedToken || !decodedToken.id) {
+		return response.status(401).json({
+			error: 'token missing or invalid',
+		});
+	}
+	const blog = await Blog.findById(request.params.id);
+	const user = await User.findById(decodedToken.id);
+
+	if (blog.user.toString() === user._id.toString()) {
+		await Blog.findByIdAndRemove(request.params.id);
+		response.status(204).end();
+	} else {
+		response.status(401).json({ error: 'Not authorized' });
+	}
 });
 
 blogsRouter.put('/:id', async (request, response) => {
